# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "My API"
    },
    "paths": {
...
        "/api/QuanLyDatVe/TaoLichChieu": {
            "post": {
                "tags": [
                    "QuanLyDatVe"
                ],
                "operationId": "TaoLichChieu",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "lich",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/LichChieuInsert"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/QuanLyRap/LayThongTinHeThongRap": {
            "get": {
                "tags": [
                    "QuanLyRap"
                ],
                "operationId": "LayThongTinHeThongRap",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maHeThongRap",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyRap/LayThongTinCumRapTheoHeThong": {
            "get": {
                "tags": [
                    "QuanLyRap"
                ],
                "operationId": "LayThongTinCumRapTheoHeThong",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maHeThongRap",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyRap/LayThongTinLichChieuHeThongRap": {
            "get": {
                "tags": [
                    "QuanLyRap"
                ],
                "operationId": "LayThongTinLichChieuHeThongRap",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "maHeThongRap",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "maNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyRap/LayThongTinLichChieuPhim": {
            "get": {
                "tags": [
                    "QuanLyRap"
                ],
                "operationId": "LayThongTinLichChieuPhim",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaPhim",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 0
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "LichChieuInsert": {
            "type": "object",
            "properties": {
                "maPhim": {
                    "format": "int32",
                    "type": "integer"
                },
                "ngayChieuGioChieu": {
                    "type": "string"
                },
                "maRap": {
                    "type": "string"
                },
                "giaVe": {
                    "format": "double",
                    "type": "number"
                }
            }
        },
...
    }
}